fix(api): improving api validations

+++ api/src/app/useCases/categories/deleteCategory/DeleteCategoryUseCase.js
@@ -1,9 +1,15 @@
+const isValidUUID = require("../../../utils/isValidUUID​");
+
 class DeleteCategoryUseCase {
   constructor(categoriesRepository) {
     this.categoriesRepository = categoriesRepository
   }
 
   async execute(id) {
+    if(!isValidUUID(category_id)) {
+      throw new AppError('Invalid category', 400);
+    }
+
     await this.categoriesRepository.delete(id);
   }
 }
